refactor: Input 컴포넌트 및 스토리북 ts 적용
