--- src/components/Atoms/Input/input.stories.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { BlockInput, UnderlineInput } from '.';
-
-export default {
-  title: 'Atoms/Input',
-};
-
-export const BlockedInputExample = args => <BlockInput {...args} />;
-
-export const UnderlineInputExample = args => <UnderlineInput {...args} />;
