+++ src/styles/theme.ts
@@ -1,10 +1,16 @@
 export const PALETTE = {
+  WHITE_001: '#FFFFFF',
+  GRAY_001: '#ECEBF1',
+  GRAY_002: '#737373',
+  GRAY_003: '#383838',
   RED_001: '#E24141',
+  RED_002: '#FF0000',
   BLUE_001: '#547CE4',
   GREEN_001: '#73BC6D',
   PINK_001: '#DE59B9',
   PINK_002: '#E76E9A',
   MINT_001: '#4CD2BA',
+  MINT_002: '#04C09E',
   ORANGE_001: '#F37D3B',
   YELLOW_001: '#FBCD58',
 };
